Some style fixes
